+++ azext_iot/commands.py
@@ -183,7 +183,7 @@ def load_command_table(self, _):
         cmd_group.command("monitor-events", "iot_central_monitor_events")
         cmd_group.command("validate-messages", "iot_central_validate_messages")
         cmd_group.command(
-            "show-provisioning-info", "iot_cental_show_provisioning_information"
+            "show-provisioning-info", "iot_central_show_provisioning_information"
         )
 
     with self.command_group(
