+++ azext_iot/operations/central.py
@@ -6,18 +6,18 @@
 
 from knack.util import CLIError
 from azext_iot._factory import _bind_sdk
-from azext_iot.common._azure import get_iot_hub_token_from_central_app_id
+from azext_iot.common.central_functions import get_iot_hub_token_from_central_app_id
 from azext_iot.common.shared import SdkType
 from azext_iot.common.utility import unpack_msrest_error, init_monitoring
 from azext_iot.common.sas_token_auth import BasicSasTokenAuthentication
-from azext_iot.common._azure import show_iot_central_provisioning_information
+from azext_iot.common.central_functions import show_iot_central_provisioning_information
 
 
 def find_between(s, start, end):
     return (s.split(start))[1].split(end)[0]
 
 
-def iot_cental_show_provisioning_information(
+def iot_central_show_provisioning_information(
     cmd, app_id, device_id=None, central_api_uri="api.azureiotcentral.com"
 ):
     show_iot_central_provisioning_information(cmd, app_id, device_id)
